requirements.txt
scikit-learn
scanpy
torch
pytorch_lightning
google-api-python-client
kedro